[Node] Possible fix for error (0x006B5E5A) while loading the map
I have noticed when loading a map in devmap sometimes the cause of it is the Node system.

Game::CL_IsCgameInitialized() only returns true when the player is actually in the game, but what if he is in the loading screen? With this clcState returns 0 when in main menu so if clcState isn't 0 we will return.